Update distribution method for better browser support
See https://github.com/fluture-js/fluture-sanctuary-types/issues/73

The main problem was the `.cjs` extensions, which confused content
delivery networks. Simply changing the extension to `.js` won't work,
because that'd confuse Node 12 and up.

So besides changing the extension, the file has also been moved into a
sub-directory containing a package.json file that tells newer Node
versions that the files here are CommonJS (without the need for a file
extension).

This commit also adds documentation for the different ways that the
module can now be obtained.